requirements.txt
fastapi
uvicorn
tensorflow-cpu==2.20.0
keras==3.13.2
opencv-python-headless
numpy==1.23.5
pillow